+++ src/main/java/com/whb/cloud/service/info/CbcUserService.java
@@ -34,7 +34,7 @@ public interface CbcUserService extends IService<CbcUserEntity> {
      * @Param: [user_money, user_id]
      * @return: int
      **/
-    int updateUserMoney(int user_money,int user_id);
+    int updateUserMoney(double user_money,int user_id);
 
     /**
      * @Author: wanghanbin
